changed lineHeight type to double

+++ lib/viltage_config.dart
@@ -5,7 +5,7 @@ class VilTAGEConfig {
   int width = 20;
   int height = 20;
   int updatePS = 15;
-  int fontSize = 14;
+  double fontSize = 14.toDouble();
   int lineHeight = 1;
   String backgroundColor = "#FFFFFF";
   ParagraphElement pe;
